Add fast inverse square Vector3.mag() method; Change physics timestep; Fix UI clicking

+++ P3DPGE/src/math/Vector3.h
@@ -187,8 +187,20 @@ inline Vector3 Vector3::cross(const Vector3& rhs) const {
 }
 
 inline float Vector3::mag() const {
-	return std::sqrtf(x * x + y * y + z * z);
+	return std::sqrt(x * x + y * y + z * z);
 }
+
+////ref: https://betterexplained.com/articles/understanding-quakes-fast-inverse-square-root/
+//inline float Vector3::mag() const {
+//	ASSERT(CHAR_BIT*sizeof(float) == 32 && CHAR_BIT*sizeof(int32) == 32, "This mag method only works if float and int are 32bit");
+//	float k = x * x + y * y + z * z;
+//	float kHalf = .5f * k;
+//	int32 i = *(int32*)&k;
+//	i = 0x5f3759df - (i >> 1);
+//	k = *(float*)&i;
+//	k = k*(1.5f - kHalf*k*k);
+//	return 1.f / k;
+//}
 
 inline void Vector3::normalize() {
 	if (*this != Vector3(0, 0, 0)) {
